% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directconnect_operations.R
\name{directconnect_describe_virtual_gateways}
\alias{directconnect_describe_virtual_gateways}
\title{Lists the virtual private gateways owned by the AWS account}
\usage{
directconnect_describe_virtual_gateways()
}
\description{
Lists the virtual private gateways owned by the AWS account.
}
\details{
You can create one or more AWS Direct Connect private virtual interfaces
linked to a virtual private gateway.
}
\keyword{internal}
